services/inference/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
ultralytics>=8.1.0
opencv-python
numpy<2.0.0
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
google-generativeai==0.8.3
huggingface-hub[inference]>=0.25.0
lapx>=0.5.5
Pillow>=10.0.0
scipy>=1.10.0
edge-tts>=6.1.9
roboflow
transformers>=4.36.0
timm>=0.9.0
SoccerNet>=0.1.62
